create categories and products views
register controller check user send id to profil & create check in redirectauthanticated middleware & create categories and products views & create routs for category and product ansd routing set on  sidbar

+++ database/migrations/2020_03_16_192816_create_orders_table.php
@@ -15,11 +15,11 @@ class CreateOrdersTable extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('product_id');
-            $table->unsignedBigInteger('qty');
+            $table->unsignedInteger('user_id');
+            $table->unsignedInteger('product_id');
+            $table->unsignedInteger('qty');
             $table->string('status');
-            $table->unsignedBigInteger('payment_id');
+            $table->unsignedInteger('payment_id');
             $table->timestamps();
             $table->softDeletes();
         });
